+++ Sources/Sentry/SentryOptions.m
@@ -130,6 +130,8 @@ NSString *const kSentryDefaultEnvironment = @"production";
         _enableSwizzling = YES;
         self.sendClientReports = YES;
         self.swiftAsyncStacktraces = NO;
+        self.enableSpotlight = NO;
+        self.spotlightUrl = @"http://localhost:8969/stream";
 
 #if TARGET_OS_OSX
         NSString *dsn = [[[NSProcessInfo processInfo] environment] objectForKey:@"SENTRY_DSN"];
@@ -488,6 +490,13 @@ NSString *const kSentryDefaultEnvironment = @"production";
     }
 #endif // SENTRY_HAS_METRIC_KIT
 
+    [self setBool:options[@"enableSpotlight"]
+            block:^(BOOL value) { self->_enableSpotlight = value; }];
+
+    if ([options[@"spotlightUrl"] isKindOfClass:[NSString class]]) {
+        self.spotlightUrl = options[@"spotlightUrl"];
+    }
+
     return YES;
 }
 
@@ -692,6 +701,17 @@ isValidSampleRate(NSNumber *sampleRate)
 }
 
 #endif // SENTRY_UIKIT_AVAILABLE
+
+- (void)setEnableSpotlight:(BOOL)value
+{
+    _enableSpotlight = value;
+#if defined(RELEASE)
+    if (value) {
+        SENTRY_LOG_WARN(@"Enabling Spotlight for a release build. We recommend running Spotlight "
+                        @"only for local development.");
+    }
+#endif // defined(RELEASE)
+}
 
 #if defined(DEBUG) || defined(TEST) || defined(TESTCI)
 - (NSString *)debugDescription