+++ Tests/SentryTests/SentryOptionsTest.m
@@ -4,6 +4,7 @@
 #import "SentrySDK.h"
 #import "SentryTests-Swift.h"
 #import <XCTest/XCTest.h>
+@import Nimble;
 
 @interface SentryOptionsTest : XCTestCase
 
@@ -560,7 +561,8 @@
         @"failedRequestStatusCodes" : [NSNull null],
         @"enableTimeToFullDisplayTracing" : [NSNull null],
         @"enableTracing" : [NSNull null],
-        @"swiftAsyncStacktraces" : [NSNull null]
+        @"swiftAsyncStacktraces" : [NSNull null],
+        @"spotlightUrl" : [NSNull null]
     }
                                                 didFailWithError:nil];
 
@@ -644,6 +646,8 @@
     XCTAssertNil(options.profilesSampleRate);
     XCTAssertNil(options.profilesSampler);
 #endif
+
+    XCTAssertTrue([options.spotlightUrl isEqualToString:@"http://localhost:8969/stream"]);
 }
 
 - (void)testSetValidDsn
@@ -1253,6 +1257,23 @@
     XCTAssertNotNil(debugDescription);
     XCTAssert([debugDescription containsString:@"sampleRate: 0.123"]);
     XCTAssert([debugDescription containsString:@"tracesSampler: <__NSGlobalBlock__: "]);
+}
+
+- (void)testEnableSpotlight
+{
+    [self testBooleanField:@"enableSpotlight" defaultValue:NO];
+}
+
+- (void)testSpotlightUrl
+{
+    SentryOptions *options = [self getValidOptions:@{ @"spotlightUrl" : @"http://localhost:1010" }];
+    XCTAssertEqualObjects(options.spotlightUrl, @"http://localhost:1010");
+
+    SentryOptions *options2 = [self getValidOptions:@{ @"spotlightUrl" : @"" }];
+    XCTAssertEqualObjects(options2.spotlightUrl, @"");
+
+    SentryOptions *options3 = [self getValidOptions:@{ @"spotlightUrl" : @2 }];
+    XCTAssertEqualObjects(options3.spotlightUrl, @"http://localhost:8969/stream");
 }
 
 #pragma mark - Private