+++ Tests/SentryTests/SentryTests-Bridging-Header.h
@@ -174,6 +174,7 @@
 #import "SentrySpan.h"
 #import "SentrySpanId.h"
 #import "SentrySpanOperations.h"
+#import "SentrySpotlightTransport.h"
 #import "SentryStacktrace.h"
 #import "SentryStacktraceBuilder.h"
 #import "SentrySubClassFinder.h"